+++ backend/StreetCatCareCommunity/src/main/java/com/streetCat/service/impl/CommentServiceImpl.java
@@ -1,10 +1,12 @@
 package com.streetCat.service.impl;
 
+import com.streetCat.dao.AdminMapper;
 import com.streetCat.dao.CommentMapper;
 import com.streetCat.pojo.Comment;
 import com.streetCat.service.CommentService;
 import com.streetCat.utils.BusinessException;
 import com.streetCat.utils.RandomUtil;
+import com.streetCat.utils.RedisCountUtil;
 import com.streetCat.vo.request.CreateCommentRequest;
 import com.streetCat.vo.response.CommentResp;
 import com.streetCat.vo.response.PageResponse;
@@ -15,18 +17,22 @@ import org.springframework.util.CollectionUtils;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class CommentServiceImpl implements CommentService {
     private final CommentMapper commentMapper;
+    private final RedisCountUtil redisCountUtil;
+    private final AdminMapper adminMapper;
     @Override
     public Comment createComment(String userId, CreateCommentRequest request) {
-        if (request.getParentId() != null && request.getParentId().isEmpty()) {
+        if (request.getParentId().isEmpty()) {
             request.setParentId(null);
         }
         String id = String.valueOf(RandomUtil.nextId());
+        redisCountUtil.incrementCommentCount(request.getTargetType(), Long.valueOf(request.getTargetId()));
         commentMapper.insertComment(id, userId, request);
         return commentMapper.selectCommentById(id);
     }
@@ -73,7 +79,10 @@ public class CommentServiceImpl implements CommentService {
         if (CollectionUtils.isEmpty(ids)) {
             return;
         }
-        // TODO: 后续补充存在性与权限校验
+        if (!adminMapper.existsById(userId) && (!Objects.equals(commentMapper.selectCommentById(ids.getFirst()).getAuthor().getId(), userId))) {
+            throw new BusinessException("你不是管理员或发布评论的本人，无权删除该评论");
+        }
+        redisCountUtil.decrementCommentCount(userId, Long.valueOf(ids.getFirst()));
         commentMapper.deleteByIds(ids);
     }
 }
